refactor 2/x [skip ci]
